requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
Pillow~=9.1.0
websocket-client~=1.3.2
click==8.1.2
loguru==0.6.0
beautifulsoup4~=4.10.0
stem~=1.8.0
numpy~=1.25.0
```